Build-Blog/backend/requirements.txt
Flask
pandas
gunicorn
python-dotenv
matplotlib
google-cloud-storage
google-cloud-bigquery
google-cloud-aiplatform
google-api-core
numpy
cloud-sql-python-connector[pymysql]
pymysql
sqlalchemy
sqlparse